refactor code to place ListItemLink component into its own file
