+++ src/App.js
@@ -7,15 +7,15 @@ import Poodle from "../src/assets/PoodleGraphic.png"
 import "./App.scss";
 /****************************** React Router *********************************** */
 import { BrowserRouter, Routes, Route, NavLink, MemoryRouter, useLocation, Link as RouterLink } from "react-router-dom";
-import { StaticRouter } from 'react-router-dom/server';
 /*************************Custom Components*************************** */
 import Player from "./Components/Player";
 import MinicastList from "./Components/MinicastList";
 import Nav from "./Components/Nav";
 import Dashboard from "./Components/Dashboard";
 import DynamicMinicast from "./Components/DynamicMinicast.js";
+import ListItemLink from "./Components/ListItemLink";
 /************************** MUI Components***************************************** */
-import ListItemIcon from '@mui/material/ListItemIcon';
+
 import {
   Box,
   Toolbar,
@@ -76,27 +76,7 @@ function App() {
       });
   }, []);
 
-  function ListItemLink(props) {
-    const { icon, primary, to, button } = props;
-
-    const renderLink = React.useMemo(
-      () =>
-        React.forwardRef(function Link(itemProps, ref) {
-          return <RouterLink to={to} ref={ref} {...itemProps} role={undefined} />;
-        }),
-      [to],
-    );
-
-    return (
-      <li>
-        <ListItem button={button} component={renderLink}>
-          {icon ? <ListItemIcon>{icon}</ListItemIcon> : null}
-          <ListItemText primary={primary} />
-        </ListItem>
-      </li>
-    );
-  }
-
+  
 
   const handleAutoPlaySwitch = event => {
     setAutoplay(event.target.checked);
